requirements.txt
# Core Streamlit App
streamlit==1.33.0

# MongoDB connection
pymongo==4.7.0

# Data handling
pandas==2.2.1
numpy==1.26.4

# Visualization
matplotlib==3.8.3

# Face Recognition / OpenCV
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80

# Utilities
Pillow==10.2.0
python-dateutil==2.9.0
pathlib==1.0.1

# File operations, hashing, JSON handling (standard libs included but added for safety)
pyyaml==6.0.1

# Serialization (pickle included in Python but added for completeness)
cloudpickle==3.0.0

# Security (password hashing is done using hashlib but extra utils)
bcrypt==4.1.2

# Additional compatibility packages
requests==2.31.0
protobuf==4.25.3
